boards: build.rs: add shared build crate
All boards can include this as a dev dependency to use the common
build.rs functions.
