+++ boards/build_scripts/Cargo.toml
@@ -1,0 +1,10 @@
+# Licensed under the Apache License, Version 2.0 or the MIT License.
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+# Copyright Tock Contributors 2024.
+
+[package]
+name = "tock_build_scripts"
+description = """Helper functions for compiling Tock boards."""
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
