+++ boards/build_scripts/tock_kernel_layout.ld
